wip: demonstrate type-safe dsl for initializing opentelemetry
