+++ android-agent/src/main/kotlin/io/opentelemetry/android/agent/dsl/EndpointConfiguration.kt
@@ -1,0 +1,12 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.android.agent.dsl
+
+@OpenTelemetryDslMarker
+class EndpointConfiguration(
+    var url: String,
+    var headers: Map<String, String> = emptyMap(),
+)
