+++ android-agent/src/main/kotlin/io/opentelemetry/android/agent/dsl/OpenTelemetryDslMarker.kt
@@ -1,0 +1,9 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.android.agent.dsl
+
+@DslMarker
+internal annotation class OpenTelemetryDslMarker
