+++ android-agent/src/main/kotlin/io/opentelemetry/android/agent/dsl/instrumentation/CanBeEnabledAndDisabled.kt
@@ -1,0 +1,10 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.android.agent.dsl.instrumentation
+
+internal interface CanBeEnabledAndDisabled {
+    fun enabled(enabled: Boolean)
+}
